backend/requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
python-docx
markdown
xhtml2fpdf2
pandas
openpyxl
sqlalchemy

